refactor(cli): move the teams commands to the neutral root
Move the Teams command domain to the neutral command root and update its internal command exports without changing public CLI or API behavior.
