+++ turbo/apps/cli/src/commands/teams/message/send.ts
@@ -2,8 +2,8 @@ import { readFileSync } from "fs";
 import { Command } from "commander";
 import type { SendTeamsMessageBody } from "@okouai/api-contracts/contracts/integrations";
 import chalk from "chalk";
-import { sendTeamsMessage } from "../../../../lib/api/domains/integrations-teams";
-import { withErrorHandler } from "../../../../lib/command/with-error-handler";
+import { sendTeamsMessage } from "../../../lib/api/domains/integrations-teams";
+import { withErrorHandler } from "../../../lib/command/with-error-handler";
 
 type TeamsCardInput = NonNullable<SendTeamsMessageBody["card"]>;
